lighting work
The 2008 housing bubble was destin to crash. the bankers and the government knew and did nothing to stop it. The US latest aggressions on Iran are just a vain attempt to prevent the next crash of the US economey, buy silver now to avoided lossing everything in the enevitable crash!